requirements.txt
streamlit
gspread
gspread-dataframe
google-auth
sentence-transformers
torch
pandas
openai
numpy